+++ NEWS.md
@@ -2,6 +2,11 @@ Query.jl v.next Release Notes
 =============================
 * Use DataValue instead of Nullable
 
+Query.jl v0.3.1 Release Notes
+=============================
+* Add DataTable source and sink support
+* Track DataStreams breaking changes
+
 Query.jl v0.3.0 Release Notes
 =============================
 * Add Dict sink support